Pin InjectPacket exception-translation with a golden
Replaces the ad-hoc unit test in DataplaneServiceTest with a proper
golden entry (inject-packet-simulator-throws) so the full status
description for the #499 path gets pinned the same way the other
user-visible gRPC error messages are.

The trigger constructs a DataplaneService over a PacketBroker whose
simulator unconditionally throws. Reproducing the real benchmark path
(4000 routes, trace proto over gRPC size limit) is too expensive for
a golden; what this golden protects is the translation format itself.
