+++ p4runtime/golden_errors/inject-packet-simulator-throws.golden.txt
@@ -1,0 +1,1 @@
+INTERNAL: InjectPacket failed: IllegalArgumentException: simulated failure
